Added packages, organized some files
