--- app/src/main/java/edu/mit/mitmobile2/shuttles/PersistRoutesInDbTask.java
@@ -1,52 +1,0 @@
-package edu.mit.mitmobile2.shuttles;
-
-import android.content.ContentValues;
-import android.os.AsyncTask;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
-
-import edu.mit.mitmobile2.DBAdapter;
-import edu.mit.mitmobile2.DatabaseObject;
-import edu.mit.mitmobile2.MitMobileApplication;
-import edu.mit.mitmobile2.Schema;
-import edu.mit.mitmobile2.shuttles.model.MITShuttleRoute;
-import timber.log.Timber;
-
-public class PersistRoutesInDbTask extends AsyncTask<List<MITShuttleRoute>, Void, Void> {
-
-    public PersistRoutesInDbTask() {
-    }
-
-    @Override
-    protected Void doInBackground(List<MITShuttleRoute>... params) {
-
-        List<DatabaseObject> updatedRoutes = new ArrayList<>();
-
-        DBAdapter dbAdapter = MitMobileApplication.dbAdapter;
-
-        // Get previous IDs
-        Set<String> ids = dbAdapter.getAllIds(Schema.Route.TABLE_NAME, Schema.Route.ALL_COLUMNS, Schema.Route.ROUTE_ID);
-
-        for (MITShuttleRoute r : params[0]) {
-            if (ids.contains(r.getId())) {
-                ContentValues values = new ContentValues();
-                r.fillInContentValues(values, dbAdapter);
-                dbAdapter.db.update(Schema.Route.TABLE_NAME, values,
-                        Schema.Route.ROUTE_ID + " = \'" + r.getId() + "\'", null);
-            } else {
-                updatedRoutes.add(r);
-            }
-        }
-
-        dbAdapter.batchPersist(updatedRoutes, Schema.Route.TABLE_NAME);
-        return null;
-    }
-
-    @Override
-    protected void onPostExecute(Void aVoid) {
-//        List<MITShuttleRoute> allRoutes = MitMobileApplication.dbAdapter.getAllRoutes();
-        Timber.d("Finished");
-    }
-}
