+++ app/src/main/java/edu/mit/mitmobile2/shuttles/activities/ShuttleRouteActivity.java
@@ -1,21 +1,22 @@
-package edu.mit.mitmobile2.shuttles;
+package edu.mit.mitmobile2.shuttles.activities;
+
+import edu.mit.mitmobile2.shuttles.callbacks.MapFragmentCallback;
+import edu.mit.mitmobile2.shuttles.fragment.ShuttleRouteFragment;
 
 import android.os.Bundle;
 
 import edu.mit.mitmobile2.MITActivity;
 import edu.mit.mitmobile2.R;
-import edu.mit.mitmobile2.shuttles.fragment.ShuttleStopFragment;
 
-public class ShuttleStopActivity extends MITActivity implements MapFragmentCallback {
+public class ShuttleRouteActivity extends MITActivity implements MapFragmentCallback {
 
-    ShuttleStopFragment fragment;
+    ShuttleRouteFragment fragment = new ShuttleRouteFragment();
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_with_frame);
 
-        fragment = new ShuttleStopFragment();
         getFragmentManager().beginTransaction().replace(R.id.fragment_frame, fragment).commit();
     }
