+++ app/src/main/java/edu/mit/mitmobile2/shuttles/activities/ShuttleStopActivity.java
@@ -1,21 +1,22 @@
-package edu.mit.mitmobile2.shuttles;
-
-import edu.mit.mitmobile2.shuttles.fragment.ShuttleRouteFragment;
+package edu.mit.mitmobile2.shuttles.activities;
 
 import android.os.Bundle;
 
 import edu.mit.mitmobile2.MITActivity;
 import edu.mit.mitmobile2.R;
+import edu.mit.mitmobile2.shuttles.callbacks.MapFragmentCallback;
+import edu.mit.mitmobile2.shuttles.fragment.ShuttleStopFragment;
 
-public class ShuttleRouteActivity extends MITActivity implements MapFragmentCallback {
+public class ShuttleStopActivity extends MITActivity implements MapFragmentCallback {
 
-    ShuttleRouteFragment fragment = new ShuttleRouteFragment();
+    ShuttleStopFragment fragment;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_with_frame);
 
+        fragment = new ShuttleStopFragment();
         getFragmentManager().beginTransaction().replace(R.id.fragment_frame, fragment).commit();
     }
