+++ app/src/main/java/edu/mit/mitmobile2/shuttles/adapter/MITShuttleAdapter.java
@@ -16,8 +16,8 @@ import java.util.List;
 import butterknife.ButterKnife;
 import butterknife.InjectView;
 import edu.mit.mitmobile2.R;
-import edu.mit.mitmobile2.shuttles.ShuttleAdapterCallback;
-import edu.mit.mitmobile2.shuttles.ShuttleUtils;
+import edu.mit.mitmobile2.shuttles.callbacks.ShuttleAdapterCallback;
+import edu.mit.mitmobile2.shuttles.utils.ShuttleUtils;
 import edu.mit.mitmobile2.shuttles.model.MITShuttleRoute;
 import edu.mit.mitmobile2.shuttles.model.MITShuttleStopWrapper;
 import edu.mit.mitmobile2.shuttles.model.MitMiniShuttleRoute;
