+++ app/src/main/java/edu/mit/mitmobile2/shuttles/callbacks/MapFragmentCallback.java
@@ -1,4 +1,4 @@
-package edu.mit.mitmobile2.shuttles;
+package edu.mit.mitmobile2.shuttles.callbacks;
 
 public interface MapFragmentCallback {
     void setActionBarTitle(String title);
