+++ app/src/main/java/edu/mit/mitmobile2/shuttles/fragment/ShuttlesFragment.java
@@ -36,10 +36,10 @@ import edu.mit.mitmobile2.R;
 import edu.mit.mitmobile2.Schema;
 import edu.mit.mitmobile2.shuttles.MITShuttlesProvider;
 import edu.mit.mitmobile2.shuttles.MitCursorLoader;
-import edu.mit.mitmobile2.shuttles.ShuttleAdapterCallback;
-import edu.mit.mitmobile2.shuttles.ShuttleRouteActivity;
-import edu.mit.mitmobile2.shuttles.ShuttleStopActivity;
-import edu.mit.mitmobile2.shuttles.ShuttlesDatabaseHelper;
+import edu.mit.mitmobile2.shuttles.callbacks.ShuttleAdapterCallback;
+import edu.mit.mitmobile2.shuttles.activities.ShuttleRouteActivity;
+import edu.mit.mitmobile2.shuttles.activities.ShuttleStopActivity;
+import edu.mit.mitmobile2.shuttles.utils.ShuttlesDatabaseHelper;
 import edu.mit.mitmobile2.shuttles.adapter.MITShuttleAdapter;
 import edu.mit.mitmobile2.shuttles.model.MitMiniShuttleRoute;
 import timber.log.Timber;
