+++ app/src/main/java/edu/mit/mitmobile2/shuttles/model/MITShuttleRoute.java
@@ -14,7 +14,7 @@ import java.util.List;
 import edu.mit.mitmobile2.DBAdapter;
 import edu.mit.mitmobile2.DatabaseObject;
 import edu.mit.mitmobile2.Schema;
-import edu.mit.mitmobile2.shuttles.ShuttlesDatabaseHelper;
+import edu.mit.mitmobile2.shuttles.utils.ShuttlesDatabaseHelper;
 
 
 public class MITShuttleRoute extends DatabaseObject implements Parcelable {
